Refactoring `osmosis.py`
