+++ src/dbt_osmosis/core/__init__.py
@@ -1,0 +1,18 @@
+# Backwards compatibility imports
+from dbt_osmosis.core.config import *  # noqa: F403
+from dbt_osmosis.core.sql_operations import *  # noqa: F403
+from dbt_osmosis.core.schema.parser import *  # noqa: F403
+from dbt_osmosis.core.schema.reader import *  # noqa: F403
+from dbt_osmosis.core.schema.writer import *  # noqa: F403
+from dbt_osmosis.core.node_filters import *  # noqa: F403
+from dbt_osmosis.core.introspection import *  # noqa: F403
+from dbt_osmosis.core.settings import *  # noqa: F403
+from dbt_osmosis.core.path_management import *  # noqa: F403
+from dbt_osmosis.core.restructuring import *  # noqa: F403
+from dbt_osmosis.core.inheritance import *  # noqa: F403
+from dbt_osmosis.core.sync_operations import *  # noqa: F403
+from dbt_osmosis.core.plugins import *  # noqa: F403
+from dbt_osmosis.core.transforms import *  # noqa: F403
+
+# Backwards compatibility - all functions available through star imports
+# No explicit __all__ needed since we use star imports from submodules
