+++ src/dbt_osmosis/core/schema/formats/__init__.py
@@ -1,0 +1,2 @@
+# Format handlers will be imported here once created
+__all__: list[str] = []
